refactor: More tidying
