+++ pubky-homeserver/src/admin_server/routes/generate_signup_token.rs
@@ -1,8 +1,8 @@
 use axum::{extract::State, http::StatusCode, response::IntoResponse, Json};
 
 use crate::{
-    data_directory::user_quota::UserQuota,
     persistence::sql::signup_code::{SignupCodeId, SignupCodeRepository},
+    persistence::user_quota::UserQuota,
     shared::{HttpError, HttpResult},
 };
 
@@ -37,7 +37,6 @@ pub async fn generate_signup_token_with_limits(
     State(state): State<AppState>,
     Json(config): Json<UserQuota>,
 ) -> HttpResult<impl IntoResponse> {
-    // Validate rate strings before touching the DB — return 422 for bad values.
     config
         .validate_rate_roundtrips()
         .map_err(|e| HttpError::new_with_message(StatusCode::UNPROCESSABLE_ENTITY, e))?;
