+++ pubky-homeserver/src/persistence/mod.rs
@@ -7,3 +7,5 @@
 
 pub mod files;
 pub mod sql;
+/// Per-user quota domain type, cache, and encoding.
+pub mod user_quota;
