+++ pubky-homeserver/src/persistence/sql/entities/user.rs
@@ -3,8 +3,8 @@ use sea_query::{Expr, Iden, PostgresQueryBuilder, Query, SimpleExpr};
 use sea_query_binder::SqlxBinder;
 use sqlx::{postgres::PgRow, FromRow, Row};
 
-use crate::data_directory::user_quota::{UserQuota, UserQuotaPatch};
 use crate::persistence::sql::UnifiedExecutor;
+use crate::persistence::user_quota::{UserQuota, UserQuotaPatch};
 
 pub const USER_TABLE: &str = "users";
 
@@ -309,7 +309,7 @@ impl UserRepository {
         pubkey: &pubky_common::crypto::PublicKey,
         quota_mb: u64,
     ) -> UserEntity {
-        use crate::data_directory::user_quota::QuotaOverride;
+        use crate::persistence::user_quota::QuotaOverride;
         let user = Self::create(pubkey, &mut db.pool().into()).await.unwrap();
         let config = UserQuota {
             storage_quota_mb: QuotaOverride::Value(quota_mb),
@@ -582,7 +582,7 @@ mod tests {
     #[pubky_test_utils::test]
     async fn test_set_quota() {
         use crate::data_directory::quota_config::BandwidthRate;
-        use crate::data_directory::user_quota::QuotaOverride;
+        use crate::persistence::user_quota::QuotaOverride;
         use std::str::FromStr;
 
         let db = SqlDb::test().await;
@@ -650,7 +650,7 @@ mod tests {
     #[test]
     fn test_limits_mixed_null_and_values() {
         use crate::data_directory::quota_config::BandwidthRate;
-        use crate::data_directory::user_quota::QuotaOverride;
+        use crate::persistence::user_quota::QuotaOverride;
         use std::str::FromStr;
 
         let user = UserEntity {
@@ -679,7 +679,7 @@ mod tests {
 
     #[test]
     fn test_limits_unlimited_values() {
-        use crate::data_directory::user_quota::QuotaOverride;
+        use crate::persistence::user_quota::QuotaOverride;
 
         let user = UserEntity {
             id: 1,
@@ -704,7 +704,7 @@ mod tests {
 
     #[test]
     fn test_limits_invalid_rate_string_treated_as_default() {
-        use crate::data_directory::user_quota::QuotaOverride;
+        use crate::persistence::user_quota::QuotaOverride;
 
         let user = UserEntity {
             id: 1,
